+++ src/App.js
@@ -3,8 +3,8 @@ import { Route, Switch } from 'react-router-dom';
 import { Container } from '@material-ui/core';
 import Header from './components/header/header';
 import ItemList from './components/item-list/item-list';
-import useNews from './useNews';
 import Title from './components/title/title';
+import useNews from './hooks/useNews';
 import SingleNewsItem from './components/single-news-item/single-news-item';
 import classes from './App.module.css';
