+++ src/hooks/use-news.js
@@ -1,11 +1,10 @@
 import { useEffect, useRef } from "react";
 import { useDispatch } from "react-redux";
-import { getNews } from "./api/api";
-import { loadNews, loadNewsFail } from "./redux/actions/actions";
-import { TIME_INTERVAL } from "./utils/utils";
+import { loadNews, loadNewsFail } from "../redux/actions/news";
+import { getNews } from "./requests/news";
+import { TIME_INTERVAL } from "../shared/constants";
 
 const useNews = () => {
-
   const dispatch = useDispatch();
   const timerRef = useRef();
