+++ src/redux/reducers/news-item.js
