+++ src/redux/reducers/news.js
