+++ src/shared/constants.js
@@ -1,0 +1,2 @@
+export const TIME_INTERVAL = 60000;
+export const ITEMS_LIMIT = 50;
