+++ src/shared/requests/item.js
@@ -1,0 +1,2 @@
+
+export const baseURL = 'https://hacker-news.firebaseio.com/v0';
