+++ src/shared/utils/comments/comments.js
@@ -1,18 +1,3 @@
-export const baseURL = 'https://hacker-news.firebaseio.com/v0';
-export const TIME_INTERVAL = 60000;
-export const ITEMS_LIMIT = 50;
-
-export function getLocalDateFormat(time){
-  if (!time) return;
-  return new Date(time * 1000).toLocaleString();
-}
-
-export function textParser(text){
-  const parseText = new DOMParser().parseFromString(text, 'text/html');
-  return parseText.body.textContent;
-}
-
-
 export const buildTree = (comments, postId) => {
   const commentsById = {};
